add functionalities and changes related to logging in

+++ routers/book.router.ts
@@ -27,6 +27,10 @@ export const bookRouter = Router()
             throw new ValidationError('There is no such book.')
         }
 
+        if(await book.countGivenBooks() > 0) {
+            throw new ValidationError('Cannot remove taken book')
+        }
+
         await book.delete();
         res.end();
     })
